--- sharding-core/sharding-core-route/src/main/java/org/apache/shardingsphere/sharding/route/type/complex/ShardingComplexRoutingEngine.java
@@ -1,83 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.sharding.route.type.complex;
-
-import com.google.common.base.Function;
-import com.google.common.base.Optional;
-import com.google.common.collect.Lists;
-import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.underlying.common.exception.ShardingSphereException;
-import org.apache.shardingsphere.sql.parser.relation.statement.SQLStatementContext;
-import org.apache.shardingsphere.sharding.route.engine.condition.ShardingConditions;
-import org.apache.shardingsphere.sharding.route.type.ShardingRouteEngine;
-import org.apache.shardingsphere.underlying.route.context.RouteResult;
-import org.apache.shardingsphere.sharding.route.type.standard.ShardingStandardRoutingEngine;
-import org.apache.shardingsphere.core.rule.BindingTableRule;
-import org.apache.shardingsphere.core.rule.ShardingRule;
-import org.apache.shardingsphere.core.rule.TableRule;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.TreeSet;
-
-/**
- * Sharding complex routing engine.
- * 
- * @author gaohongtao
- * @author zhangliang
- */
-@RequiredArgsConstructor
-public final class ShardingComplexRoutingEngine implements ShardingRouteEngine {
-    
-    private final Collection<String> logicTables;
-    
-    private final SQLStatementContext sqlStatementContext;
-    
-    private final ShardingConditions shardingConditions;
-    
-    @Override
-    public RouteResult route(final ShardingRule shardingRule) {
-        Collection<RouteResult> result = new ArrayList<>(logicTables.size());
-        Collection<String> bindingTableNames = new TreeSet<>(String.CASE_INSENSITIVE_ORDER);
-        for (String each : logicTables) {
-            Optional<TableRule> tableRule = shardingRule.findTableRule(each);
-            if (tableRule.isPresent()) {
-                if (!bindingTableNames.contains(each)) {
-                    result.add(new ShardingStandardRoutingEngine(tableRule.get().getLogicTable(), sqlStatementContext, shardingConditions).route(shardingRule));
-                }
-                Optional<BindingTableRule> bindingTableRule = shardingRule.findBindingTableRule(each);
-                if (bindingTableRule.isPresent()) {
-                    bindingTableNames.addAll(Lists.transform(bindingTableRule.get().getTableRules(), new Function<TableRule, String>() {
-                        
-                        @Override
-                        public String apply(final TableRule input) {
-                            return input.getLogicTable();
-                        }
-                    }));
-                }
-            }
-        }
-        if (result.isEmpty()) {
-            throw new ShardingSphereException("Cannot find table rule and default data source with logic tables: '%s'", logicTables);
-        }
-        if (1 == result.size()) {
-            return result.iterator().next();
-        }
-        return new ShardingCartesianRoutingEngine(result).route(shardingRule);
-    }
-}
